+++ src/contexts/KeyboardContext.tsx
@@ -1,12 +1,13 @@
+import en from '../translation/en'
 import * as React from 'react'
 import type { TextStyle, ViewStyle } from 'react-native'
-import { defaultKeyboardContext, defaultKeyboardValues } from './KeyboardProvider'
-import type {
+import {
   CategoryTranslation,
   EmojiType,
   CategoryTypes,
   CategoryPosition,
   EmojisByCategory,
+  CATEGORIES,
 } from '../types'
 
 export type OnEmojiSelected = (emoji: EmojiType) => void
@@ -55,6 +56,51 @@ export type ContextValues = {
   searchPhrase: string
   setSearchPhrase: (phrase: string) => void
   renderList: EmojisByCategory[]
+}
+
+export const defaultKeyboardContext: Required<KeyboardProps> = {
+  open: false,
+  onClose: () => {},
+  onEmojiSelected: (_emoji: EmojiType) => {},
+  emojiSize: 28,
+  containerStyles: {},
+  knobStyles: {},
+  headerStyles: {},
+  expandable: true,
+  hideHeader: false,
+  defaultHeight: '40%',
+  expandedHeight: '80%',
+  backdropColor: '#00000055',
+  categoryColor: '#000000',
+  activeCategoryColor: '#005b96',
+  categoryContainerColor: '#e3dbcd',
+  activeCategoryContainerColor: '#ffffff',
+  onCategoryChangeFailed: (info) => {
+    console.warn(info)
+  },
+  translation: en,
+  disabledCategories: [],
+  enableRecentlyUsed: false,
+  categoryPosition: 'floating',
+  enableSearchBar: false,
+  closeSearchColor: '#00000055',
+  searchBarStyles: {},
+  searchBarTextStyles: {},
+  searchBarPlaceholderColor: '#00000055',
+  categoryOrder: [...CATEGORIES],
+  onRequestClose: () => {},
+  categoryContainerStyles: {},
+  disableSafeArea: false,
+}
+
+export const defaultKeyboardValues: ContextValues = {
+  activeCategoryIndex: 0,
+  setActiveCategoryIndex: () => {},
+  numberOfColumns: 5,
+  width: 0,
+  searchPhrase: '',
+  setSearchPhrase: (_phrase: string) => {},
+  renderList: [],
 }
 
 export const KeyboardContext = React.createContext<Required<KeyboardProps> & ContextValues>({
